Add util/update_code: code/src = src/stubs + src/tests
If modifying stubs or tests, this updates the skeleton sources in
code/src to match. It also will *replace* the tests on any existing
source files if desired, so an existing implementation can be tested
with the latest test cases.

This should make it easier to maintain the skeletons and to improve the
test cases, ensuring everything is in sync.