chunk_type: safe-to-copy bit is 0 for critical chunks
The PNG spec states "the safe-to-copy bit will always be 0 for critical
chunks". All of the tests should respect this. One test has been added
to ensure this (chunk_type::test_invalid_safetocopy).

This required changing many tests, as they used "RuSt", which is not
a valid chunk type.

Fixes #11.